tests: instantiate QGIS object for every test case
some TestCases may fiddle with the QGIS objects (for example adding it to a
QMainWindow which may delete the underlying C++ object in the end.

setup a fresh QGIS environment for every TestCase by moving the setup code
into the class setup. Also check in get_qgis_app() if stuff is defined but
C++ objects are missing. Don't know if this is effective though.