changes in travis test script to actually let build fail if tests failed
the current script does not fail if tests fail because the "set" command
at the beginning was wrong.

also we now want to actually handle the test script results internally, so
an instant fail is now disabled, we print a nice message and then return
with the exit code of the testrunner script.