do not convert geodetic coordinates to absolute values (fixes #12 - backport)
Geographic coordinates need to be converted to absolute values because
we display hemispheric direction as button labels. On the other hand
negative values are valid and important for some geodetic CRS, for example
EPSG:31254 (MGI / Austria GK West). This distinction was not made and all
coordinates were converted to absolute values on clicking the canvas or
changing the input CRS.